+++ service/src/main/kotlin/traversium/moderation/ModerationServiceApplication.kt
@@ -1,9 +1,10 @@
 package traversium.moderation
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.kafka.KafkaAutoConfiguration
 import org.springframework.boot.runApplication
 
-@SpringBootApplication
+@SpringBootApplication(exclude = [KafkaAutoConfiguration::class])
 class ModerationServiceApplication
 
 fun main(args: Array<String>) {